requirements.txt
# Remote Desktop - Python Dependencies

# Core dependencies
websockets>=12.0
Pillow>=10.0.0
numpy>=1.24.0
pygame>=2.5.0
pyautogui>=0.9.54

# macOS screen capture (optional but recommended on macOS)
pyobjc-framework-Quartz>=9.0; sys_platform == 'darwin'

# Development/Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
